debugged user creration, but need to debug passport local. That's where passing data us stuck.
