+++ routes/api/auth.js
@@ -2,10 +2,13 @@ const express = require('express');
 const router = express.Router(); 
 const validator = require('validator');
 const passport = require('passport');
-
+// const passportLocal = require('../../passport/passport-signup');
 
 router.post('/', (req, res, next) => {
+    console.log('in Auth Signup Router')
+    console.log(req.body);
     const validationResult = validateSignupForm(req.body);
+    console.log('validation result: ' + validationResult);
     if (!validationResult.success) {
         return res.status(400).json({
         success: false,
@@ -15,6 +18,7 @@ router.post('/', (req, res, next) => {
     }
 
     return passport.authenticate('local-signup', (err) => {
+        console.log('in return passport authenticate');
         if (err) {
         if (err.name === 'MongoError' && err.code === 11000) {
             // the 11000 Mongo code is for a duplication email error
@@ -42,6 +46,8 @@ router.post('/', (req, res, next) => {
 });
 
 function validateSignupForm(payload) {
+    console.log('in payload');
+    console.log(payload);
     const errors = {};
     let isFormValid = true;
     let message = '';
@@ -56,15 +62,16 @@ function validateSignupForm(payload) {
         errors.password = 'Password must have at least 8 characters.';
     }
 
-    if (!payload || typeof payload.name !== 'string' || payload.name.trim().length === 0) {
-        isFormValid = false;
-        errors.name = 'Please provide your name.';
-    }
+    // if (!payload || typeof payload.firstName !== 'string' || payload.firstName.trim().length === 0) {
+    //     isFormValid = false;
+    //     errors.name = 'Please provide your name.';
+    // }
 
     if (!isFormValid) {
         message = 'Check the form for errors.';
     }
-
+    console.log('===== valid form results ======');
+    console.log(errors, isFormValid, message);
     return {
         success: isFormValid,
         message,
