+++ routes/api/petbreedinfo.js
@@ -4,6 +4,9 @@ const router = express.Router();
 const petbreedController = require('../../controllers/petbreedController');
 
 router.route('/')
+    .get(petbreedController.breedInfoRecord);
+
+router.route('/all')
     .get(petbreedController.findallBreedInfo);
 
 module.exports = router;
